requirements.txt
alabaster==0.7.8
Babel==2.3.4
colorama==0.3.7
coverage==4.1
defusedxml==0.4.1
Django==1.9.6
django-allauth==0.25.2
django-rest-auth==0.7.0
djangorestframework==3.3.3
docutils==0.12
imagesize==0.7.1
Jinja2==2.8
MarkupSafe==0.23
oauthlib==1.1.1
Pillow==3.2.0
Pygments==2.1.3
python3-openid==3.0.10
pytz==2016.4
requests==2.10.0
requests-oauthlib==0.6.1
selenium==2.53.2
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.1
sphinx-rtd-theme==0.1.9
